pyproject.toml
[project]
name = "mydeep"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "polars>=1.35.2",
    "pyarrow>=22.0.0",
    "pyspark>=4.0.1",
    "pyyaml>=6.0.3",
    "scipy>=1.16.3",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
]

[[tool.uv.index]]
url = "https://bytedpypi.byted.org/simple"
